requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
appnope==0.1.3
astroid==2.13.5
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
bidict==0.22.1
black==22.12.0
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==3.2.0
click==8.1.5
cryptography==41.0.2
dataclasses-json==0.5.12
decorator==5.1.1
dill==0.3.6
distlib==0.3.7
dnspython==2.4.0
docx2txt==0.8
eventlet==0.33.3
exceptiongroup==1.1.2
executing==1.2.0
filelock==3.12.2
Flask==2.3.2
Flask-SocketIO==5.3.4
frozenlist==1.4.0
fsspec==2023.6.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
identify==2.5.24
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
ipython==8.10.0
isort==5.12.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.1
langchain==0.0.235
langsmith==0.0.7
lazy-object-proxy==1.9.0
llama-index==0.7.9
MarkupSafe==2.1.3
marshmallow==3.19.0
matplotlib-inline==0.1.6
mccabe==0.7.0
multidict==6.0.4
mypy==0.991
mypy-extensions==1.0.0
nest-asyncio==1.5.6
nltk==3.8.1
nodeenv==1.8.0
numexpr==2.8.4
numpy==1.25.1
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==2.0.3
parso==0.8.3
pathspec==0.11.1
pexpect==4.8.0
pickleshare==0.7.5
pip==23.2
platformdirs==3.9.1
pluggy==1.2.0
pre-commit==3.2.0
prompt-toolkit==3.0.39
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==1.10.11
pydantic_core==2.3.0
Pygments==2.15.1
pylint==2.15.10
pypdf==3.12.2
pytest==7.2.1
pytest-asyncio==0.21.0
pytest-dotenv==0.5.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.5.1
python-socketio==5.8.0
pytz==2023.3
PyYAML==6.0
rake-nltk==1.0.6
regex==2023.6.3
requests==2.31.0
ruff==0.0.259
setuptools==49.2.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.19
stack-data==0.6.2
tenacity==8.2.2
tiktoken==0.4.0
tokenize-rt==5.1.0
tomli==2.0.1
tomlkit==0.11.8
tqdm==4.65.0
traitlets==5.9.0
types-docutils==0.20.0.1
types-pyOpenSSL==23.2.0.1
types-redis==4.5.5.0
types-requests==2.28.11.8
types-setuptools==67.1.0.0
types-urllib3==1.26.25.13
typing_extensions==4.7.1
typing-inspect==0.9.0
tzdata==2023.3
urllib3==1.26.16
virtualenv==20.24.0
wcwidth==0.2.6
Werkzeug==2.3.6
wrapt==1.15.0
yarl==1.9.2
zipp==3.16.2
python-decouple==3.5

